requirements.txt
unstructured
tiktoken
langchain
pypdf
openai
langchain
pandas
numpy
python-dotenv
